pyproject.toml
[project]
name = "dask-climate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dask[complete]>=2025.2.0",
    "matplotlib>=3.10.1",
    "netcdf4>=1.7.2",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "xarray>=2025.1.2",
]
